# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Nixyah API - Food marketplace for Côte d'Ivoire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List recent stories
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                required: [stories]

    post:
      operationId: createStory
      tags: [stories]
      summary: Create a story (chef only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoryInput"
      responses:
        "201":
          description: Story created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stories/{id}:
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                required: [orders]

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Story:
      type: object
      properties:
        id:
          type: string
        chefId:
          type: string
        chefName:
          type: string
        chefCoverColor:
          type: string
        caption:
          type: string
        dishName:
          type: string
        price:
          type: number
        emoji:
          type: string
        bgColor:
          type: string
        createdAt:
          type: string
        expiresAt:
          type: string
      required: [id, chefId, chefName, chefCoverColor, caption, createdAt, expiresAt]
    CreateStoryInput:
      type: object
      properties:
        caption:
          type: string
        dishName:
          type: string
        price:
          type: number
        emoji:
          type: string
        bgColor:
          type: string
      required: [caption]
    Order:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        chefId:
          type: string
        chefName:
          type: string
        status:
          type: string
          enum: [pending, accepted, preparing, ready, delivered]
        total:
          type: number
        occasion:
          type: string
        persons:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [id, clientId, chefId, chefName, status, total, createdAt]
    OrderItem:
      type: object
      properties:
        dishId:
          type: string
        dishName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required: [dishId, dishName, quantity, price]
    CreateOrderInput:
      type: object
      properties:
        chefId:
          type: string
        occasion:
          type: string
        persons:
          type: integer
        budget:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [chefId]
...