# lib/api-zod/src/generated/types/createStoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nixyah API - Food marketplace for Côte d'Ivoire
 * OpenAPI spec version: 0.2.0
 */

export interface CreateStoryInput {
  caption: string;
  dishName?: string;
  price?: number;
  emoji?: string;
  bgColor?: string;
}