# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nixyah API - Food marketplace for Côte d'Ivoire
 * OpenAPI spec version: 0.2.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  clientId: string;
  chefId: string;
  chefName: string;
  status: OrderStatus;
  total: number;
  occasion?: string;
  persons?: number;
  notes?: string;
  createdAt: string;
  items?: OrderItem[];
}